requirements.txt
python-telegram-bot
vk_api
python-dotenv
